+++ saleor/graphql/account/mutations/base.py
@@ -7,6 +7,8 @@ from django.utils.http import urlsafe_base64_encode
 from graphql_jwt.exceptions import PermissionDenied
 
 from ....account import emails, events as account_events, models
+from ....dashboard.emails import send_set_password_customer_email
+from ...account.i18n import I18nMixin
 from ...account.types import Address, AddressInput, User
 from ...core.mutations import (
     BaseMutation,
@@ -16,6 +18,8 @@ from ...core.mutations import (
     UpdateMetaBaseMutation,
 )
 
+BILLING_ADDRESS_FIELD = "default_billing_address"
+SHIPPING_ADDRESS_FIELD = "default_shipping_address"
 INVALID_TOKEN = "Invalid or expired token."
 
 
@@ -186,6 +190,87 @@ class BaseAddressDelete(ModelDeleteMutation):
 
         response.user = user
         return response
+
+
+class UserInput(graphene.InputObjectType):
+    first_name = graphene.String(description="Given name.")
+    last_name = graphene.String(description="Family name.")
+    email = graphene.String(description="The unique email address of the user.")
+    is_active = graphene.Boolean(required=False, description="User account is active.")
+    note = graphene.String(description="A note about the user.")
+
+
+class UserAddressInput(graphene.InputObjectType):
+    default_billing_address = AddressInput(
+        description="Billing address of the customer."
+    )
+    default_shipping_address = AddressInput(
+        description="Shipping address of the customer."
+    )
+
+
+class CustomerInput(UserInput, UserAddressInput):
+    pass
+
+
+class UserCreateInput(CustomerInput):
+    send_password_email = graphene.Boolean(
+        description="Send an email with a link to set a password"
+    )
+
+
+class BaseCustomerCreate(ModelMutation, I18nMixin):
+    """Base mutation for customer create used by staff and account."""
+
+    class Arguments:
+        input = UserCreateInput(
+            description="Fields required to create a customer.", required=True
+        )
+
+    class Meta:
+        abstract = True
+
+    @classmethod
+    def clean_input(cls, info, instance, data):
+        shipping_address_data = data.pop(SHIPPING_ADDRESS_FIELD, None)
+        billing_address_data = data.pop(BILLING_ADDRESS_FIELD, None)
+        cleaned_input = super().clean_input(info, instance, data)
+
+        if shipping_address_data:
+            shipping_address = cls.validate_address(
+                shipping_address_data,
+                instance=getattr(instance, SHIPPING_ADDRESS_FIELD),
+            )
+            cleaned_input[SHIPPING_ADDRESS_FIELD] = shipping_address
+
+        if billing_address_data:
+            billing_address = cls.validate_address(
+                billing_address_data, instance=getattr(instance, BILLING_ADDRESS_FIELD)
+            )
+            cleaned_input[BILLING_ADDRESS_FIELD] = billing_address
+        return cleaned_input
+
+    @classmethod
+    def save(cls, info, instance, cleaned_input):
+        # FIXME: save address in user.addresses as well
+        default_shipping_address = cleaned_input.get(SHIPPING_ADDRESS_FIELD)
+        if default_shipping_address:
+            default_shipping_address.save()
+            instance.default_shipping_address = default_shipping_address
+        default_billing_address = cleaned_input.get(BILLING_ADDRESS_FIELD)
+        if default_billing_address:
+            default_billing_address.save()
+            instance.default_billing_address = default_billing_address
+
+        is_creation = instance.pk is None
+        super().save(info, instance, cleaned_input)
+
+        # The instance is a new object in db, create an event
+        if is_creation:
+            account_events.customer_account_created_event(user=instance)
+
+        if cleaned_input.get("send_password_email"):
+            send_set_password_customer_email.delay(instance.pk)
 
 
 class UserUpdateMeta(UpdateMetaBaseMutation):
